成功调用embeddings

+++ deploy.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-cp config-example.json config.json
+if [ ! -f config.json ]
+then
+	cp config-example.json config.json
+fi
 pip install -r requirements.txt
-python3 webui.py
+nohup python3 -u webui.py > webui.log 2>&1
